Move set importer to /admin/ytg/import; /admin/import-set becomes a redirect
git mv with a minimal diff: drop the page's own <TopNav/> and client-side
permission redirect (the ytg layout now gates server-side), neutralize the
chrome wrappers the shell provides, demote its h1. All importer logic is
untouched; the /api/admin/import-set route stays where it is.
